# backend/archive/unreferenced_examples/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Affine Formula Screening",
    "version": "1.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/catalog": {
      "get": {
        "summary": "Catalog",
        "operationId": "catalog_api_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/examples": {
      "get": {
        "summary": "Examples",
        "operationId": "examples_api_examples_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/context": {
      "post": {
        "summary": "Context",
        "operationId": "context_api_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Formula"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments": {
      "post": {
        "summary": "Assessment",
        "operationId": "assessment_api_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Formula"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare": {
      "post": {
        "summary": "Comparison",
        "operationId": "comparison_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comparison"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Comparison": {
        "properties": {
          "previous": {
            "$ref": "#/components/schemas/Formula"
          },
          "current": {
            "$ref": "#/components/schemas/Formula"
          }
        },
        "type": "object",
        "required": [
          "previous",
          "current"
        ],
        "title": "Comparison"
      },
      "Formula": {
        "properties": {
          "formula_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Formula Id"
          },
          "version_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Version Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "product_category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Product Category"
          },
          "age_group": {
            "type": "string",
            "enum": [
              "under_12",
              "12_and_above"
            ],
            "title": "Age Group"
          },
          "regions": {
            "items": {
              "type": "string",
              "enum": [
                "US",
                "EU",
                "UK",
                "CA"
              ]
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Regions"
          },
          "physical_form": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Physical Form"
          },
          "intended_use": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Intended Use"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ingredients"
          },
          "generate_explanation": {
            "type": "boolean",
            "title": "Generate Explanation",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "formula_id",
          "version_id",
          "name",
          "product_category",
          "age_group",
          "regions",
          "physical_form",
          "intended_use",
          "ingredients"
        ],
        "title": "Formula"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Ingredient": {
        "properties": {
          "material_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Material Id"
          },
          "concentration_percent": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Concentration Percent"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "material_id",
          "concentration_percent"
        ],
        "title": "Ingredient"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}